refactor: 제목 수정 모달 공통화
